Splitted ut-s by server.
